[backend] Fix

+++ openbas-api/src/main/java/io/openbas/executors/caldera/service/CalderaExecutorService.java
@@ -145,6 +145,7 @@ public class CalderaExecutorService implements Runnable {
               this.agentService.createOrUpdateAgent(agentToUpdate);
             } else {
               // New agent to create for the endpoint
+              agent.setAsset(endpointToUpdate);
               this.agentService.createOrUpdateAgent(agent);
             }
           } else {
@@ -153,24 +154,26 @@ public class CalderaExecutorService implements Runnable {
             this.agentService.createOrUpdateAgent(agent);
           }
         } else {
-          Optional<io.openbas.database.model.Agent> optionalAgent =
-              this.agentService.getAgentByAgentDetailsForAnAsset(
-                  optionalEndpoint.get().getId(),
-                  agent.getExecutedByUser(),
-                  agent.getDeploymentMode(),
-                  agent.getPrivilege(),
-                  CALDERA_EXECUTOR_TYPE);
-          if (optionalAgent.isPresent()) {
-            io.openbas.database.model.Agent existingAgent = optionalAgent.get();
-            if ((now().toEpochMilli() - agent.getLastSeen().toEpochMilli()) > DELETE_TTL) {
-              log.info(
-                  "Found stale endpoint "
-                      + endpoint.getName()
-                      + ", deleting the agent "
-                      + existingAgent.getExecutedByUser()
-                      + " in it...");
-              this.client.deleteAgent(existingAgent);
-              this.agentService.deleteAgent(existingAgent.getId());
+          if (optionalEndpoint.isPresent()) {
+            Optional<io.openbas.database.model.Agent> optionalAgent =
+                    this.agentService.getAgentByAgentDetailsForAnAsset(
+                            optionalEndpoint.get().getId(),
+                            agent.getExecutedByUser(),
+                            agent.getDeploymentMode(),
+                            agent.getPrivilege(),
+                            CALDERA_EXECUTOR_TYPE);
+            if (optionalAgent.isPresent()) {
+              io.openbas.database.model.Agent existingAgent = optionalAgent.get();
+              if ((now().toEpochMilli() - agent.getLastSeen().toEpochMilli()) > DELETE_TTL) {
+                log.info(
+                        "Found stale endpoint "
+                                + endpoint.getName()
+                                + ", deleting the agent "
+                                + existingAgent.getExecutedByUser()
+                                + " in it...");
+                this.client.deleteAgent(existingAgent);
+                this.agentService.deleteAgent(existingAgent.getId());
+              }
             }
           }
         }
